backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prevencion-api"
version = "1.0.0"
description = "PrevencionApp — Telemedicina preventiva para detección de inflamación sinovial"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "gunicorn>=21.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "supabase>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "httpx>=0.27.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-storage>=2.16.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "Pillow>=10.3.0",
    "onnxruntime>=1.18.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
